Folders

+++ react-files/src/Components/ArchiveObj.js
@@ -73,6 +73,9 @@ class ArchiveObj extends Component {
     // delete project content
     delete() {
         let _name = this.state.name.toString();
+        if (window.confirm("למחוק את הקבצים של פרויקט \"" + _name + "\" ?") == false)
+            return;
+
         this.deleteFolderContents(_name);
         console.log("Document '" + _name + "' files deleted!");
         this.state.getArchive();
